api/urls.py
from django.urls import path, include
from django.contrib.auth.views import LogoutView
from rest_framework.routers import DefaultRouter
from .views import TaskViewSet

router = DefaultRouter()
router.register(r'tasks', TaskViewSet, basename='task')

urlpatterns = [
  path('list/', TaskViewSet.as_view({'get': 'task_list'}), name='api_task_list'),
]

