requirements.txt
Django==3.2.6
djangorestframework==3.12.4
psycopg2==2.9.1
django-simple-history==3.4.0